Upgraded packages added Seq logginh Updated LlamaSharp usages to conform with v0.6

+++ src/Semantic.Kernel.LlamaSharp/Tests/Semantic.Kernel.LlamaSharp.Tests/Tests/TestOfLlamaSharpSemanticKernel.cs
@@ -69,7 +69,7 @@ public class TestOfLlamaSharpSemanticKernel : IAsyncLifetime
         // Load weights into memory
         var parameters = new ModelParams(modelPath)
         {
-            Seed = RandomNumberGenerator.GetInt32(int.MaxValue),
+            Seed = (uint)RandomNumberGenerator.GetInt32(int.MaxValue),
         };
         using var model = LLamaWeights.LoadFromFile(parameters);
         var ex = new StatelessExecutor(model, parameters);
