+++ src/Semantic.Kernel.LlamaSharp/Tests/Semantic.Kernel.LlamaSharp.Tests/Tests/TestOfLlamaSharpSemanticKernelChat.cs
@@ -83,7 +83,7 @@ public class TestOfLlamaSharpSemanticKernelChat : IAsyncLifetime
         // Load weights into memory
         var parameters = new ModelParams(modelPath)
         {
-            Seed = RandomNumberGenerator.GetInt32(int.MaxValue),
+            Seed = (uint)RandomNumberGenerator.GetInt32(int.MaxValue),
         };
         using var model = LLamaWeights.LoadFromFile(parameters);
         using var context = model.CreateContext(parameters);
